moving duplicate classes to different path in csv module to fix android compilation
