+++ dataframe-csv/src/main/kotlin/org/jetbrains/kotlinx/dataframe/io/readDelimStr.kt
@@ -2,24 +2,24 @@ package org.jetbrains.kotlinx.dataframe.io
 
 import org.jetbrains.kotlinx.dataframe.DataFrame
 import org.jetbrains.kotlinx.dataframe.api.ParserOptions
-import org.jetbrains.kotlinx.dataframe.documentation.CommonReadDelimDocs
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.ADJUST_CSV_SPECS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.ALLOW_MISSING_COLUMNS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.COL_TYPES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.DELIM_DELIMITER
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.FIXED_COLUMN_WIDTHS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.HAS_FIXED_WIDTH_COLUMNS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.HEADER
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.IGNORE_EMPTY_LINES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.IGNORE_EXCESS_COLUMNS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.IGNORE_SURROUNDING_SPACES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.PARSER_OPTIONS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.PARSE_PARALLEL
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.QUOTE
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.READ_LINES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.SKIP_LINES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.TEXT_READ
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.TRIM_INSIDE_QUOTED
+import org.jetbrains.kotlinx.dataframe.documentationCsv.CommonReadDelimDocs
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.ADJUST_CSV_SPECS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.ALLOW_MISSING_COLUMNS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.COL_TYPES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.DELIM_DELIMITER
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.FIXED_COLUMN_WIDTHS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.HAS_FIXED_WIDTH_COLUMNS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.HEADER
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.IGNORE_EMPTY_LINES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.IGNORE_EXCESS_COLUMNS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.IGNORE_SURROUNDING_SPACES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.PARSER_OPTIONS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.PARSE_PARALLEL
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.QUOTE
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.READ_LINES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.SKIP_LINES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.TEXT_READ
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.TRIM_INSIDE_QUOTED
 import org.jetbrains.kotlinx.dataframe.impl.io.readDelimImpl
 
 /**
