+++ dataframe-csv/src/main/kotlin/org/jetbrains/kotlinx/dataframe/io/readTsv.kt
@@ -4,29 +4,29 @@ package org.jetbrains.kotlinx.dataframe.io
 
 import org.jetbrains.kotlinx.dataframe.DataFrame
 import org.jetbrains.kotlinx.dataframe.api.ParserOptions
-import org.jetbrains.kotlinx.dataframe.documentation.CommonReadDelimDocs
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.ADJUST_CSV_SPECS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.ALLOW_MISSING_COLUMNS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.COL_TYPES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.COMPRESSION
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.FILE_OR_URL_READ
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.FILE_READ
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.FIXED_COLUMN_WIDTHS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.HAS_FIXED_WIDTH_COLUMNS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.HEADER
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.IGNORE_EMPTY_LINES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.IGNORE_EXCESS_COLUMNS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.IGNORE_SURROUNDING_SPACES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.INPUT_STREAM_READ
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.PARSER_OPTIONS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.PARSE_PARALLEL
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.PATH_READ
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.QUOTE
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.READ_LINES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.SKIP_LINES
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.TRIM_INSIDE_QUOTED
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.TSV_DELIMITER
+import org.jetbrains.kotlinx.dataframe.documentationCsv.CommonReadDelimDocs
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.ADJUST_CSV_SPECS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.ALLOW_MISSING_COLUMNS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.COL_TYPES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.COMPRESSION
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.FILE_OR_URL_READ
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.FILE_READ
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.FIXED_COLUMN_WIDTHS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.HAS_FIXED_WIDTH_COLUMNS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.HEADER
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.IGNORE_EMPTY_LINES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.IGNORE_EXCESS_COLUMNS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.IGNORE_SURROUNDING_SPACES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.INPUT_STREAM_READ
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.PARSER_OPTIONS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.PARSE_PARALLEL
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.PATH_READ
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.QUOTE
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.READ_LINES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.SKIP_LINES
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.TRIM_INSIDE_QUOTED
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.TSV_DELIMITER
 import org.jetbrains.kotlinx.dataframe.impl.io.readDelimImpl
 import java.io.File
 import java.io.FileInputStream
