+++ dataframe-csv/src/main/kotlin/org/jetbrains/kotlinx/dataframe/io/writeCsv.kt
@@ -3,19 +3,19 @@
 package org.jetbrains.kotlinx.dataframe.io
 
 import org.jetbrains.kotlinx.dataframe.AnyFrame
-import org.jetbrains.kotlinx.dataframe.documentation.CommonWriteDelimDocs
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.ADJUST_CSV_FORMAT
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.COMMENT_CHAR
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.CSV_DELIMITER
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.ESCAPE_CHAR
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.FILE_WRITE
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.HEADER_COMMENTS
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.INCLUDE_HEADER
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.PATH_WRITE
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.QUOTE
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.QUOTE_MODE
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.RECORD_SEPARATOR
-import org.jetbrains.kotlinx.dataframe.documentation.DelimParams.WRITER_WRITE
+import org.jetbrains.kotlinx.dataframe.documentationCsv.CommonWriteDelimDocs
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.ADJUST_CSV_FORMAT
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.COMMENT_CHAR
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.CSV_DELIMITER
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.ESCAPE_CHAR
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.FILE_WRITE
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.HEADER_COMMENTS
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.INCLUDE_HEADER
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.PATH_WRITE
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.QUOTE
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.QUOTE_MODE
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.RECORD_SEPARATOR
+import org.jetbrains.kotlinx.dataframe.documentationCsv.DelimParams.WRITER_WRITE
 import org.jetbrains.kotlinx.dataframe.impl.io.writeDelimImpl
 import java.io.File
 import java.io.FileWriter
